Read DXTPOSIX; Output to Parquet
Removed skip DXT_POSIX option as the log files may now
be read as intended using the experimental branch of Darshan.
(Further information will hopefully be provided in README using
a link pointed to the Docker image I made to do this.)

Completely changed internals of how the record were processed:
- removed assumptions that certain modules (e.g. LUSTRE) were
	singletons.
- put everything into classes so they can take care of
	writing to parquet files themselves according to
	whether they were counter, fcounter, or read/write
	dataframe collections.
- added record names to the output files to deconvolute
	LUSTRE, POSIX, etc. module files from different
	records.

A few assumptions of uniqueness were made. I added checks
and ValueErrors to hopefully crash in the case these
assumptions prove false.

Note that currently MPI-IO and DXT_MPIIO are not currently
processed, but it should be easy enough to add. For now
it will just vomit it to output, so it will be obvious
when this is an issue.