Reduce all outputs to 1 file per module; Add tests
All modules are output as a single DF, keyed by juid and
jobid of the report it came from.

Added a test along with a set of example data from the
XPDaMoDa repository. This test ensures that all data
in the independent DXT_POSIX files was correctly condensed
into a single file.